21.03 pull request

+++ code/user_list_indexer/src/com/turning_leaf_technologies/reindexer/UserListIndexer.java
@@ -1,5 +1,6 @@
 package com.turning_leaf_technologies.reindexer;
 
+import com.turning_leaf_technologies.encryption.EncryptionUtils;
 import com.turning_leaf_technologies.indexing.IndexingUtils;
 import com.turning_leaf_technologies.indexing.Scope;
 import org.apache.logging.log4j.Logger;
@@ -35,13 +36,15 @@ class UserListIndexer {
 	private HashSet<Long> usersThatCanShareLists = new HashSet<>();
 	private SolrClient openArchivesServer;
 	private PreparedStatement getListDisplayNameAndAuthorStmt;
+	private final String serverName;
 
-	UserListIndexer(Ini configIni, Connection dbConn, Logger logger){
+	UserListIndexer(String serverName, Ini configIni, Connection dbConn, Logger logger){
+		this.serverName = serverName;
 		this.dbConn = dbConn;
 		this.logger = logger;
 		//Load a list of all list publishers
 		try {
-			PreparedStatement listPublishersStmt = dbConn.prepareStatement("SELECT userId FROM `user_roles` INNER JOIN roles on user_roles.roleId = roles.roleId where name = 'listPublisher'");
+			PreparedStatement listPublishersStmt = dbConn.prepareStatement("SELECT userId FROM user_roles INNER JOIN roles on user_roles.roleId = roles.roleId inner join role_permissions on role_permissions.roleId = roles.roleId where permissionId = (select id from permissions where name = 'Include Lists In Search Results')");
 			ResultSet listPublishersRS = listPublishersStmt.executeQuery();
 			while (listPublishersRS.next()){
 				usersThatCanShareLists.add(listPublishersRS.getLong(1));
@@ -66,7 +69,6 @@ class UserListIndexer {
 		groupedWorkServer = groupedWorkHttpBuilder.build();
 		HttpSolrClient.Builder openArchivesHttpBuilder = new HttpSolrClient.Builder("http://localhost:" + solrPort + "/solr/open_archives");
 		openArchivesServer = openArchivesHttpBuilder.build();
-
 
 		scopes = IndexingUtils.loadScopes(dbConn, logger);
 	}
@@ -105,7 +107,7 @@ class UserListIndexer {
 				//Get a list of all public lists
 				numListsStmt = dbConn.prepareStatement("select count(id) as numLists from user_list WHERE deleted = 0 AND public = 1 and searchable = 1");
 				//noinspection SpellCheckingInspection
-				listsStmt = dbConn.prepareStatement("SELECT user_list.id as id, deleted, public, searchable, title, description, user_list.created, dateUpdated, username, firstname, lastname, displayName, homeLocationId, user_id from user_list INNER JOIN user on user_id = user.id WHERE public = 1 AND deleted = 0");
+				listsStmt = dbConn.prepareStatement("SELECT user_list.id as id, deleted, public, searchable, title, description, user_list.created, dateUpdated, username, firstname, lastname, displayName, homeLocationId, user_id from user_list INNER JOIN user on user_id = user.id WHERE public = 1 AND searchable = 1 AND deleted = 0");
 			}else{
 				//Get a list of all lists that are were changed since the last update
 				//Have to process all lists because one could have been deleted, made private, or made non searchable.
@@ -138,7 +140,7 @@ class UserListIndexer {
 			}
 
 			while (allPublicListsRS.next()){
-				if (updateSolrForList(updateServer, getTitlesForListStmt, allPublicListsRS, lastReindexTime, logEntry)){
+				if (updateSolrForList(fullReindex, updateServer, getTitlesForListStmt, allPublicListsRS, lastReindexTime, logEntry)){
 					numListsIndexed++;
 				}
 				numListsProcessed++;
@@ -154,7 +156,7 @@ class UserListIndexer {
 		return numListsProcessed;
 	}
 
-	private boolean updateSolrForList(ConcurrentUpdateSolrClient updateServer, PreparedStatement getTitlesForListStmt, ResultSet allPublicListsRS, long lastReindexTime, ListIndexingLogEntry logEntry) throws SQLException, SolrServerException, IOException {
+	private boolean updateSolrForList(boolean fullReindex, ConcurrentUpdateSolrClient updateServer, PreparedStatement getTitlesForListStmt, ResultSet allPublicListsRS, long lastReindexTime, ListIndexingLogEntry logEntry) throws SQLException, SolrServerException, IOException {
 		UserListSolr userListSolr = new UserListSolr(this);
 		long listId = allPublicListsRS.getLong("id");
 
@@ -163,7 +165,7 @@ class UserListIndexer {
 		int isSearchable = allPublicListsRS.getInt("searchable");
 		long userId = allPublicListsRS.getLong("user_id");
 		boolean indexed = false;
-		if (deleted == 1 || isPublic == 0 || isSearchable == 0){
+		if (!fullReindex && (deleted == 1 || isPublic == 0 || isSearchable == 0)){
 			updateServer.deleteByQuery("id:" + listId);
 			logEntry.incDeleted();
 		}else{
@@ -174,11 +176,11 @@ class UserListIndexer {
 			long created = allPublicListsRS.getLong("created");
 			userListSolr.setCreated(created);
 
-			String displayName = allPublicListsRS.getString("displayName");
+			String displayName = EncryptionUtils.decryptString(allPublicListsRS.getString("displayName"), serverName, logEntry);
 			//noinspection SpellCheckingInspection
-			String firstName = allPublicListsRS.getString("firstname");
+			String firstName = EncryptionUtils.decryptString(allPublicListsRS.getString("firstname"), serverName, logEntry);
 			//noinspection SpellCheckingInspection
-			String lastName = allPublicListsRS.getString("lastname");
+			String lastName = EncryptionUtils.decryptString(allPublicListsRS.getString("lastname"), serverName, logEntry);
 			String userName = allPublicListsRS.getString("username");
 
 			if (userName.equalsIgnoreCase("nyt_user")) {
@@ -202,7 +204,7 @@ class UserListIndexer {
 			if (librariesByHomeLocation.containsKey(patronHomeLibrary)){
 				userListSolr.setOwningLibrary(librariesByHomeLocation.get(patronHomeLibrary));
 			} else {
-				//Don't know the owning library for some reason
+				//Don't know the owning library for some reason, most likely this is an admin user.
 				userListSolr.setOwningLibrary(-1);
 			}
 
@@ -254,7 +256,8 @@ class UserListIndexer {
 						getListDisplayNameAndAuthorStmt.setString(1, sourceId);
 						ResultSet listDisplayNameAndAuthorRS = getListDisplayNameAndAuthorStmt.executeQuery();
 						if (listDisplayNameAndAuthorRS.next()){
-							userListSolr.addListTitle("lists", sourceId, listDisplayNameAndAuthorRS.getString("title"), listDisplayNameAndAuthorRS.getString("displayName"));
+							String decryptedName = EncryptionUtils.decryptString(listDisplayNameAndAuthorRS.getString("displayName"), serverName, logEntry);
+							userListSolr.addListTitle("lists", sourceId, listDisplayNameAndAuthorRS.getString("title"), decryptedName);
 						}
 						listDisplayNameAndAuthorRS.close();
 					}else{
@@ -268,7 +271,7 @@ class UserListIndexer {
 				// Index in the solr catalog
 				SolrInputDocument document = userListSolr.getSolrDocument();
 				if (document != null){
-					updateServer.add(userListSolr.getSolrDocument());
+					updateServer.add(document);
 					if (created > lastReindexTime){
 						logEntry.incAdded();
 					}else{
